# packages/api-client/src/generated/models/IntegrationListResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { Integration } from './Integration';
export type IntegrationListResponse = {
    items: Array<Integration>;
};