# packages/api-client/src/generated/models/IntegrationTestResult.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type IntegrationTestResult = {
    ok: boolean;
    status_code?: number;
    error?: string;
};